requirements.txt
argparse
confluent-kafka
hop-client
adc
boto3